config

+++ libs/database/src/lib/entities/campaign.entity.ts
@@ -57,7 +57,7 @@ export class CampaignEntity {
   @Column({ name: 'expirationTimestamp', nullable: true })
   expireAt?: Date;
 
-  @Column('tinyint', {
+  @Column('smallint', {
     default: 0,
   })
   discountPercent!: number;
